+++ repositories/credit_repository.py
@@ -1,46 +1,53 @@
 from core.database import SessionLocal, Credit
+from models.database_schema import CreditSchema
 from repositories.base_repository import BaseRepository
 
 
-class CreditRepository(BaseRepository[Credit]):
-    def get_all(self) -> list[Credit]|iter[Credit]:
+class CreditRepository(BaseRepository[CreditSchema]):
+    def get_all(self) -> list[CreditSchema]|iter[CreditSchema]:
         try:
             with SessionLocal() as session:
                 entities = session.query(Credit).all()
-            return entities
+            return [CreditSchema.model_validate(entity) for entity in entities]
         except:
             return []
 
-    def get(self, id: int) -> Credit:
+    def get(self, id: int) -> CreditSchema:
         try:
             with SessionLocal() as session:
                 entity = session.get(Credit, id)
-            return entity
+            return CreditSchema.model_validate(entity) if entity else None
         except:
             return None
 
-    def delete(self, entity: Credit) -> bool:
+    def delete(self, id: int) -> bool:
         try:
             with SessionLocal() as session:
-                session.delete(entity)
-                session.commit()
+                entity = session.get(Credit, id)
+                if entity is not None:
+                    session.delete(entity)
+                    session.commit()
+                else:
+                    return False
             return True
         except:
             return False
 
-    def update(self, entity: Credit) -> bool:
+    def update(self, entity: CreditSchema) -> bool:
         try:
             with SessionLocal() as session:
-                session.merge(entity)
+                db_credit = Credit(**entity.model_dump())
+                session.merge(db_credit)
                 session.commit()
             return True
         except:
             return False
     
-    def insert(self, entity: Credit) -> bool:
+    def insert(self, entity: CreditSchema) -> bool:
         try:
             with SessionLocal() as session:
-                session.add(entity)
+                db_credit = Credit(**entity.model_dump())
+                session.add(db_credit)
                 session.commit()
             return True
         except:
